Fixed bugs

+++ controllers/productController.js
@@ -4,6 +4,8 @@ const Feature = require("../utils/features");
 const User = require("../models/userModel");
 const { tryCatch } = require("../middleware/asyncError");
 const ErrorHandler = require("../utils/errorHandler");
+const { validationResult } = require("express-validator");
+const constants = require("../config/constants");
 
 //get all products
 exports.getProducts = tryCatch(async (req, res, next) => {
@@ -16,16 +18,14 @@ exports.getProducts = tryCatch(async (req, res, next) => {
     .skip((page - 1) * perPage)
     .limit(perPage);
 
-  // console.log(req.userId);
-
-  const feature = new Feature(products, req.query.sortByPrice, req.query.name);
+  const feature = new Feature(products, req.query);
   feature.sort();
   const filteredProducts = feature.search();
 
   if (filteredProducts.length == 0) {
     return new ErrorHandler();
   }
-  res.status(200).json({filteredProducts});
+  res.status(constants.OK).json({ filteredProducts });
 });
 
 //get single product
@@ -35,28 +35,38 @@ exports.getProduct = tryCatch(async (req, res, next) => {
   const product = await Product.findById({ _id: productId });
 
   if (!product) {
-    return new ErrorHandler("Product not found", 404);
+    return new ErrorHandler("Product not found", constants.NOT_FOUND);
   }
 
-  res.status(200).json({ product });
+  res.status(constants.OK).json({ product });
 });
 // create a product
 exports.postProduct = tryCatch(async (req, res, next) => {
+  const errors = validationResult(req);
+  
+  if(errors.array().length>0){
+    throw new ErrorHandler(errors.array()[0].msg, constants.UNPROCESSED_ENTITY );
+  }
+  
   const name = req.body.name;
+  if(!req.file){
+     throw new ErrorHandler("Please add a product image", constants.UNPROCESSED_ENTITY);
+  }
   const imagePath = req.file.path;
+  
   const price = req.body.price;
   const description = req.body.description;
-
-  console.log(imagePath);
+  
   const product = new Product({
     name,
     imagePath,
     price,
     description,
+    creator: req.userId,
   });
 
   await product.save();
-  res.status(200).json({
+  res.status(constants.OK).json({
     message: "Product created",
     product,
   });
@@ -67,32 +77,44 @@ exports.deleteProduct = tryCatch(async (req, res, next) => {
   const productId = req.params.id;
 
   const product = await Product.findById({ _id: productId });
-  console.log(product);
+
   if (!product) {
-    return new ErrorHandler("Invalid Id", 404);
+    throw new ErrorHandler("Invalid Id", constants.BAD_REQUEST);
+  }
+
+  if (product.creator.toString() != req.userId.toString()) {
+    throw new ErrorHandler("Only creator of the product can delete", constants.UNAUTHORISED );
   }
 
   await Product.deleteOne({ _id: productId });
   await deleteImage(product.imagePath);
 
-  res.status(200).json({
-    message: "Product deleted",
+  res.status(constants.OK).json({
+    message: "Product deleted Successfully",
   });
 });
 
 // update a product
 exports.updateProduct = tryCatch(async (req, res, next) => {
+
+  const errors = validationResult(req);
+  
+  console.log(errors.array()[0]);
+  if(errors.array().length>0){
+    throw new ErrorHandler(errors.array()[0].msg, constants.UNPROCESSED_ENTITY );
+  }
+  
   const productId = req.params.id;
 
   const product = await Product.findById({ _id: productId });
 
   if (!product) {
-    return new ErrorHandler("Invalid Id", 404);
+    return new ErrorHandler("Invalid Id", constants.BAD_REQUEST);
   }
 
   const name = req.body.name ?? product.name;
   const imagePath = req.file.path ?? product.imagePath;
-  if (req.file.path) {
+  if (req.file.path && product.imagePath) {
     deleteImage(product.imagePath);
   }
   const price = req.body.price ?? product.price;
@@ -105,7 +127,7 @@ exports.updateProduct = tryCatch(async (req, res, next) => {
     { name, imagePath, price, description }
   );
 
-  res.status(200).json({
+  res.status(constants.OK).json({
     message: "Product updated successfully",
     updatedProduct,
   });
@@ -114,15 +136,14 @@ exports.updateProduct = tryCatch(async (req, res, next) => {
 // Add product to cart
 exports.addToCart = tryCatch(async (req, res, next) => {
   const productId = req.params.id;
-  if (!req.userId) {
-    return new ErrorHandler("Login is first to add to cart", 401);
+  const product = Product.findById({ _id: productId });
+  if (!product) {
+    throw new ErrorHandler("Invalid Id", constants.BAD_REQUEST);
   }
 
-  const newCartItem = {
-    product: productId,
-    quantity: 1,
-  };
-  const user = await User.findById(req.userId);
+  const userId = req.userId;
+
+  const user = await User.findById({ _id: userId });
 
   const isAddedToCart = await user.cart.find(
     (cart) => cart.product.toString() === productId.toString()
@@ -134,12 +155,12 @@ exports.addToCart = tryCatch(async (req, res, next) => {
         (cart.quantity = cart.quantity + 1), (cart.product = productId);
     });
   } else {
-    user.cart.push(newCartItem);
+    user.cart.push({ product: productId, quantity: 1 });
   }
   await user.save();
 
-  res.status(200).json({
+  res.status(constants.OK).json({
     message: "Successfully added to cart",
-    product,
+    user,
   });
 });
